+++ lib/fluent/plugin/out_tai64n_parser.rb
@@ -1,11 +1,19 @@
+require_relative 'tai64n_parser'
+
 module Fluent
   class Tai64nParserOutput < Output
     include Fluent::HandleTagNameMixin
+    include Fluent::Plugin::Tai64nParser
     Fluent::Plugin.register_output('tai64n_parser', self)
 
     # Define `log` method for v0.10.42 or earlier
     unless method_defined?(:log)
       define_method("log") { $log }
+    end
+
+    # Define `router` method of v0.12 to support v0.10 or earlier
+    unless method_defined?(:router)
+      define_method("router") { Fluent::Engine }
     end
 
     config_param :key, :string, :default => 'tai64n'
@@ -36,7 +44,7 @@ module Fluent
       es.each {|time,record|
         t = tag.dup
         filter_record(t, time, record)
-        Engine.emit(t, time, record)
+        router.emit(t, time, record)
       }
       chain.next
     end
@@ -44,33 +52,12 @@ module Fluent
     def filter_record(tag, time, record)
       begin
         record[output_key] = replace_tai64n(record[key])
-      rescue => error
-        log.warn("out_tai64n_parser: #{error.class} #{error.message} #{error.backtrace.first}")
+      rescue => e
+        log.warn("out_tai64n_parser: #{e.class} #{e.message}")
+        log.warn_backtrace
       end
-      super(tag, time, record)
+      super(tag, time, record) # HandleTagNameMixin
     end
 
-    def replace_tai64n(str)
-      tai64n, rest = str[0,25], str[25..-1]
-      parsed = parse_tai64n(tai64n)
-      if parsed
-        "#{parsed}#{rest}"
-      else
-        log.info("out_tai64n_parser: record['#{key}']='#{str}' does not start with valid tai64n")
-        str
-      end
-    end
-
-    def parse_tai64n(tai64n)
-      # @4000000052f88ea32489532c
-      # 0123456789012345678901234
-      # 0         1         2
-      #   |-------------||------|
-      return nil unless tai64n[0,2] == '@4'
-      ts = tai64n[2,15].hex
-      tf = tai64n[17,8].hex
-      t = Time.at(ts-10,tf/1000.0)
-      t.strftime("%Y-%m-%d %X.%9N")
-    end
   end
 end
